requirements.txt
quart
discord.py
